+++ src/utils.rs
@@ -3,10 +3,10 @@
 extern crate rand;
 
 //Check if number is prime or composite, and return boolean
-fn is_prime(number: usize) -> bool {
-    let root: i32 = (number as f64).sqrt() as i32 + 1;
+fn is_prime(number: u32) -> bool {
+    let root: u32 = (number as f64).sqrt() as u32 + 1;
     for i in 2..root {
-        if number % (i as usize) == 0 {
+        if number % (i as u32) == 0 {
             return false;
         }
     }
@@ -14,7 +14,7 @@ fn is_prime(number: usize) -> bool {
 }
 
 //Check if number is prime or composite, and produce output
-pub fn check_prime(number: usize) {
+pub fn check_prime(number: u32) {
     if is_prime(number) {
         println!("{} is a prime", number);
     } else {
@@ -23,7 +23,7 @@ pub fn check_prime(number: usize) {
 }
 
 //Find the closest prime above number
-pub fn prime_next(mut number: usize) {
+pub fn prime_next(mut number: u32) {
     let backup = number;
     number = number + 1;
     while is_prime(number) != true {
@@ -33,7 +33,7 @@ pub fn prime_next(mut number: usize) {
 }
 
 //Find the closest prime below number
-pub fn prime_prev(mut number: usize) {
+pub fn prime_prev(mut number: u32) {
     let backup = number;
     number = number - 1;
     while is_prime(number) != true {
@@ -43,8 +43,8 @@ pub fn prime_prev(mut number: usize) {
 }
 
 //Output the prime factors of number
-pub fn get_factors(mut number: usize) {
-    let mut factors: Vec<usize> = Vec::new();
+pub fn get_factors(mut number: u32) {
+    let mut factors: Vec<u32> = Vec::new();
     while number != 1 {
         //(number + 1) is used, as the uppermost item is excluded
         for i in 2..(number + 1) {
@@ -57,8 +57,8 @@ pub fn get_factors(mut number: usize) {
             }
             //Else check if number is divisible by i, if yes, add it
             //to the list of factors and divide with it. Repeat.
-            if number % (i as usize) == 0 {
-                number = number / (i as usize);
+            if number % (i as u32) == 0 {
+                number = (number / (i as u32)) as u32;
                 factors.push(i);
                 break;
             }
@@ -69,7 +69,7 @@ pub fn get_factors(mut number: usize) {
     factors.push(0);
     //Factor checking - Create output string from factor array
     let mut output = String::from("Prime factorization:");
-    let mut prev: usize = 0;
+    let mut prev: u32 = 0;
     let mut factor_exp = 1;
     //Always check if the current number is the same as the previous - if yes,
     //increment exponent, shift prev, and repeat. If no, concat number (prev) and
@@ -98,7 +98,7 @@ pub fn get_factors(mut number: usize) {
 }
 
 //Generate a random prime number between arguments
-pub fn generate_prime(min: usize, max: usize) {
+pub fn generate_prime(min: u32, max: u32) {
     //Error checking: Correct parameters
     if min >= max {
         println!("Error: --min value should be lower than --max value");
@@ -119,9 +119,9 @@ pub fn generate_prime(min: usize, max: usize) {
             return;
         }
     }
-    let mut number: usize = 4;
-    while !is_prime(number as usize) || number < 2 {
-        number = rand::random::<usize>() % (max - min) + min;
+    let mut number: u32 = 4;
+    while !is_prime(number as u32) || number < 2 {
+        number = rand::random::<u32>() % (max - min) + min;
     }
     println!("Generating a random prime number: {}", number);
 }
